backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.1.1
Pillow==10.1.0
pytesseract==0.3.10
easyocr==1.6.2
torch==2.2.0
# torch-2.9.0+cpu torchaudio-2.9.0+cpu torchvision-0.24.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
googletrans==4.0.0-rc1
